requirements.txt
mediapipe==0.10.20
opencv-python==4.8.1.78
numpy==1.26.3
torch==2.2.0
torchvision==0.17.0
ffmpeg-python==0.2.0
openai-whisper==20231117
transformers==4.36.2
tqdm==4.66.1
scipy==1.12.0
face-alignment==1.4.1
moviepy==1.0.3
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
soundfile==0.12.1
datasets>=2.16.1
sentencepiece>=0.1.99